requirements.txt
beautifulsoup4==4.13.4
langchain==0.3.25
langchain_core==0.3.65
langchain_google_genai==2.1.5
langgraph==0.4.8
loguru==0.7.3
mysql_connector_repackaged==0.3.1
neo4j==5.28.1
pandas==2.3.0
pydantic==2.11.7
python-dotenv==1.1.0
qdrant_client==1.14.3
Requests==2.32.4
sentence_transformers==4.1.0
tqdm==4.67.1
